% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv13_poisson_test.R
\name{cv13_poisson_test}
\alias{cv13_poisson_test}
\title{Poisson Goodness-of-Fit Test}
\usage{
cv13_poisson_test(data, residue_col)
}
\arguments{
\item{data}{A dataframe containing the observed data.}

\item{residue_col}{A string specifying the column in `data` to be tested.}
}
\value{
A dataframe object representing the chi-squared statistic and the p-value from the goodness-of-fit test.
}
\description{
Conducts a goodness-of-fit test to evaluate if the Mic data follows a Poisson distribution.
}
\examples{
# Assuming Eq_Mic is your dataframe and Mic is the column to be tested
 cv13_poisson_test(data=Eq_Mic, residue_col="Mic")
}
\author{
Chan, Mohamed, Lou, Wendy, Yang, Xiande [xiande.yang at gmail.com]
}
