% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv15_mic_mutate.R
\name{cv15_mic_mutate}
\alias{cv15_mic_mutate}
\title{Calculate Mic Statistics}
\usage{
cv15_mic_mutate(data, cleaning_event_col, residue_col)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{cleaning_event_col}{The name of the column that identifies the cleaning event.}

\item{residue_col}{The name of the column containing residue measurements.}
}
\value{
A dataframe with new columns for mean, median, grand mean, and grand median of Mic values.
}
\description{
Calculate Mic Statistics
}
\examples{
cv15_mic_mutate(data=Eq_Mic, cleaning_event_col="CleaningEvent", residue_col="Mic")
}
\author{
Chan, Mohamed, Lou, Wendy, Yang, Xiande [xiande.yang at gmail.com]
}
