% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv16_u_chart.R
\name{cv16_u_chart}
\alias{cv16_u_chart}
\title{Create a u-Chart for Poisson-distributed Data}
\usage{
cv16_u_chart(data, residue_col, cleaning_event_col)
}
\arguments{
\item{data}{A data frame containing the data for plotting.}

\item{residue_col}{The name of the column representing residue data (numeric).}

\item{cleaning_event_col}{The name of the column representing cleaning events (factor or character).}
}
\value{
A ggplot object representing the u-chart.
}
\description{
This function generates a u-chart for visualizing the stability and
capability of a process based on a Poisson distribution.
}
\examples{
cv16_u_chart(data = Eq_Mic, residue_col = "Mic", cleaning_event_col = "CleaningEvent")
}
\author{
Chan, Mohamed, Lou, Wendy, Yang, Xiande [xiande.yang at gmail.com]
}
