% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq_DAR.R
\docType{data}
\name{Eq_DAR}
\alias{Eq_DAR}
\title{Equipment Cleaning Data for DAR}
\format{
A data frame with 60 rows and 3 variables:
\describe{
  \item{DAR}{Numeric vector with DAR measurements.}
  \item{USL}{Numeric vector with Upper Specification Limits for DAR.}
  \item{CleaningEvent}{Factor vector with Cleaning Event identifiers.}
  \item{Classification}{Character vector with the deviation status for each cleaning event. Defaults to "normal".}
  \item{LIMSProjectID}{Integer or character vector with unique project IDs assigned to each row.}
}
}
\source{
Details about the data source.
}
\usage{
Eq_DAR
}
\description{
A dataset containing cleaning validation data for equipment DAR.
}
\keyword{datasets}
