% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq_Mic.R
\docType{data}
\name{Eq_Mic}
\alias{Eq_Mic}
\title{Equipment Cleaning Data for Microbial Bioburden}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{Mic}{Numeric vector with Mic measurements.}
  \item{USL}{Numeric vector with Upper Specification Limits for Mic.}
  \item{CleaningEvent}{Factor vector with Cleaning Event identifiers.}
  \item{Classification}{Character vector with the deviation status for each cleaning event. Defaults to "normal".}
  \item{LIMSProjectID}{Integer or character vector with unique project IDs assigned to each row.}
}
}
\source{
Details about the data source.
}
\usage{
Eq_Mic
}
\description{
A dataset containing cleaning validation data for microbial bioburden (Mic).
}
\keyword{datasets}
